@@ -1,79 +1,128 @@
+import 'dart:html';
+import 'dart:js_util';
+
 import 'package:flutter/material.dart';
-import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:provider/provider.dart';
 
-import 'common/Global.dart';
-import 'common/locale_model.dart';
-import 'common/theme_model.dart';
-import 'common/user_model.dart';
-
-import './routes/home_page.dart';
-import './routes/login_page.dart';
-import './routes/language_route.dart';
-import './routes/theme_change.dart';
-import 'l10n/localization_intl.dart';
-
-void main() => Global.init().then((e) => runApp(const MyApp()));
+void main() {
+  runApp(const MyApp());
+}
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
+
+  // This widget is the root of your application.
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Test',
+      theme: ThemeData(
+        // This is the theme of your application.
+        //
+        // Try running your application with "flutter run". You'll see the
+        // application has a blue toolbar. Then, without quitting the app, try
+        // changing the primarySwatch below to Colors.green and then invoke
+        // "hot reload" (press "r" in the console where you ran "flutter run",
+        // or simply save your changes to "hot reload" in a Flutter IDE).
+        // Notice that the counter didn't reset back to zero; the application
+        // is not restarted.
+        primarySwatch: Colors.green,
+      ),
+      home: const MyHomePage(title: 'Flutter Demo Home Page'),
+    );
+  }
+}
+
+class MyHomePage extends StatefulWidget {
+  const MyHomePage({super.key, required this.title});
+
+  // This widget is the home page of your application. It is stateful, meaning
+  // that it has a State object (defined below) that contains fields that affect
+  // how it looks.
+
+  // This class is the configuration for the state. It holds the values (in this
+  // case the title) provided by the parent (in this case the App widget) and
+  // used by the build method of the State. Fields in a Widget subclass are
+  // always marked "final".
+
+  final String title;
+
+  @override
+  State<MyHomePage> createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+  int _counter = 0;
+
+  void _incrementCounter() {
+    setState(() {
+      // This call to setState tells the Flutter framework that something has
+      // changed in this State, which causes it to rerun the build method below
+      // so that the display can reflect the updated values. If we changed
+      // _counter without calling setState(), then the build method would not be
+      // called again, and so nothing would appear to happen.
+      _counter++;
+    });
+  }
 
   @override
   Widget build(BuildContext context) {
-    return MultiProvider(
-      providers: [
-        ChangeNotifierProvider(create: (_) => ThemeModel()),
-        ChangeNotifierProvider(create: (_) => UserModel()),
-        ChangeNotifierProvider(create: (_) => LocaleModel()),
-      ],
-      child: Consumer2<ThemeModel, LocaleModel>(
-        builder: (BuildContext context, themeModel, localeModel, child) {
-          return MaterialApp(
-            theme: ThemeData(
-              primarySwatch: themeModel.theme,
-            ),
-            onGenerateTitle: (context) {
-              return GmLocalizations.of(context)!.title;
-            },
-            home: const HomeRoute(),
-            locale: localeModel.getLocale(),
-            //我们只支持美国英语和中文简体
-            supportedLocales: const [
-              Locale('en', 'US'), // 美国英语
-              Locale('zh', 'CN'), // 中文简体
-              //其它Locales
-            ],
-            localizationsDelegates: const [
-              // 本地化的代理类
-              GlobalMaterialLocalizations.delegate,
-              GlobalWidgetsLocalizations.delegate,
-              GmLocalizationsDelegate()
-            ],
-            localeResolutionCallback: (localeTemp, supportedLocales) {
-              if (localeModel.getLocale() != null) {
-                //如果已经选定语言，则不跟随系统
-                return localeModel.getLocale();
-              } else {
-                //跟随系统
-                Locale locale;
-                if (supportedLocales.contains(localeTemp)) {
-                  locale = localeTemp!;
-                } else {
-                  //如果系统语言不是中文简体或美国英语，则默认使用美国英语
-                  locale = const Locale('en', 'US');
-                }
-                return locale;
-              }
-            },
-            // 注册路由表
-            routes: <String, WidgetBuilder>{
-              "login": (context) => const LoginRoute(),
-              "themes": (context) => const ThemeChangeRoute(),
-              "language": (context) => const LanguageRoute(),
-            },
-          );
-        },
+    // This method is rerun every time setState is called, for instance as done
+    // by the _incrementCounter method above.
+    //
+    // The Flutter framework has been optimized to make rerunning build methods
+    // fast, so that you can just rebuild anything that needs updating rather
+    // than having to individually change instances of widgets.
+    return Scaffold(
+      appBar: AppBar(
+        // Here we take the value from the MyHomePage object that was created by
+        // the App.build method, and use it to set our appbar title.
+        title: Text(widget.title),
       ),
+      body: Center(
+        // Center is a layout widget. It takes a single child and positions it
+        // in the middle of the parent.
+        child: Column(
+          // Column is also a layout widget. It takes a list of children and
+          // arranges them vertically. By default, it sizes itself to fit its
+          // children horizontally, and tries to be as tall as its parent.
+          //
+          // Invoke "debug painting" (press "p" in the console, choose the
+          // "Toggle Debug Paint" action from the Flutter Inspector in Android
+          // Studio, or the "Toggle Debug Paint" command in Visual Studio Code)
+          // to see the wireframe for each widget.
+          //
+          // Column has various properties to control how it sizes itself and
+          // how it positions its children. Here we use mainAxisAlignment to
+          // center the children vertically; the main axis here is the vertical
+          // axis because Columns are vertical (the cross axis would be
+          // horizontal).
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            const Text(
+              'You have pushed the button this many times:',
+            ),
+            Text(
+              '$_counter',
+              style: Theme.of(context).textTheme.headlineMedium,
+            ),
+            const InputChip(label: Text('Person')),
+            ConstrainedBox(constraints: const BoxConstraints(
+              minWidth: 80.0,
+              minHeight: 50.0
+            ),
+            child: const SizedBox(
+              height: 80.0,
+              width: 80.0,
+              child: DecoratedBox(decoration: BoxDecoration(color: Colors.green)),
+            ),)
+          ],
+        ),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: _incrementCounter,
+        tooltip: 'Increment',
+        child: const Icon(Icons.add),
+      ), // This trailing comma makes auto-formatting nicer for build methods.
     );
   }
 }
