@@ -1,7 +1,6 @@
 import 'package:json_annotation/json_annotation.dart';
 import "user.dart";
 import "cacheConfig.dart";
-
 part 'profile.g.dart';
 
 @JsonSerializable()
@@ -10,13 +9,11 @@ class Profile {
 
   User? user;
   String? token;
-  num theme = 1;
+  late num theme;
   CacheConfig? cache;
   String? lastLogin;
   String? locale;
-
-  factory Profile.fromJson(Map<String, dynamic> json) =>
-      _$ProfileFromJson(json);
-
+  
+  factory Profile.fromJson(Map<String,dynamic> json) => _$ProfileFromJson(json);
   Map<String, dynamic> toJson() => _$ProfileToJson(this);
 }
